refactoring: Extract prometheus.py
